+++ RELEASE_NOTES.md
@@ -1,4 +1,5 @@
 - Added support for enum types as vertex and instance attributes
+- Implemented download / upload for PixImage as 3D slices
 
 ### 5.6.1
 - [GeometryPool] culling shader workaround (FShade write bug)